Import CoreGraphics when building for iOS
Several types that are required by FastCoding are not included in
Foundation on iOS:

- `CGFloat`
- `CGPoint`
- `CGSize`
- `CGRect`
- `CGAffineTransform`

I believe the current FastCoder examples compile because the default
.pch for iOS applications includes UIKit. I'm attempting to include
FastCoder in a static library and it will not build. This change will
perform a module import of CoreGraphics when building for iOS to resolve
compile time errors.